% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shipmentAmountDaily.R
\name{shipmentAmountDailyUI_bottom}
\alias{shipmentAmountDailyUI_bottom}
\title{生成区域}
\usage{
shipmentAmountDailyUI_bottom()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
shipmentAmountDailyUI_bottom()
}
