% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shipmentAmountDaily.R
\name{shipmentAmountDailyUI_left}
\alias{shipmentAmountDailyUI_left}
\title{请输入文件}
\usage{
shipmentAmountDailyUI_left()
}
\value{
返回值
}
\description{
请输入文件
}
\examples{
shipmentAmountDailyUI_left()
}
