% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shipmentAmountWeekly.R
\name{shipmentAmountWeeklyUI_bottom}
\alias{shipmentAmountWeeklyUI_bottom}
\title{生成区域}
\usage{
shipmentAmountWeeklyUI_bottom()
}
\value{
返回值
}
\description{
生成区域
}
\examples{
shipmentAmountWeeklyUI_bottom()
}
