% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shipmentAmountWeekly.R
\name{shipmentAmountWeeklyUI_left}
\alias{shipmentAmountWeeklyUI_left}
\title{请输入文件}
\usage{
shipmentAmountWeeklyUI_left()
}
\value{
返回值
}
\description{
请输入文件
}
\examples{
shipmentAmountWeeklyUI_left()
}
